# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FarmacoGraph API
  version: 1.0.0
  description: |
    **The API is the product.** FarmacoGraph never exposes Neo4j or PostgreSQL directly.

    Universal interface for medical education apps, simulators, flashcard generators,
    AI assistants, and third-party developers.

    ## Content Layers
    - `biomedical` — validated facts
    - `education` — pedagogical content
    - `learning` — prerequisite knowledge graph

    ## Response Modes
    - `minimal` — IDs and labels only
    - `summary` — default
    - `full` — all properties
    - `graph` — subgraph projection
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: FarmacoGraph
    url: https://farmacograph.org
servers:
  - url: /api/v1
    description: API v1
...
paths:
  /auth/token:
    post:
      tags: [Auth]
      summary: Issue access and refresh tokens
      operationId: issueToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: Token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /explain:
    get:
      tags: [Explain]
      summary: Structured reasoning chain
      operationId: explain
      parameters:
        - name: drug
          in: query
          required: true
          schema: { type: string }
        - name: effect
          in: query
          schema: { type: string }
        - name: question_type
          in: query
          schema:
            type: string
            enum: [mechanism, interaction, indication, monitoring, prerequisite]
      responses:
        "200":
          description: Explainability response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainResponse"
        "404":
          description: No path in knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /search:
    get:
      tags: [Search]
      summary: Full-text entity search
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2 }
        - name: types
          in: query
          schema:
            type: array
            items: { type: string }
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
...
components:
  parameters:
...
    PageLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    TokenRequest:
      type: object
      required: [grant_type]
      properties:
        grant_type:
          type: string
          enum: [password, api_key]
        username:
          type: string
          description: Email address for password grant
        password:
          type: string
          format: password
        api_key:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, description: Access token lifetime in seconds }
        scopes:
          type: array
          items: { type: string }
        email: { type: string, nullable: true }
        name: { type: string, nullable: true }
    ResponseMeta:
      type: object
      required: [dataset_version, api_version]
      properties:
        dataset_version: { type: string, example: "2026.1.0" }
        ontology_version: { type: string, example: "1.0.0" }
        query_time_ms: { type: integer }
        content_layers:
          type: array
          items:
            type: string
            enum: [biomedical, education, learning]
        language: { type: string, default: en }
        api_version: { type: string, example: v1 }
    EntitySummary:
      type: object
      required: [id, type, slug, label]
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        slug: { type: string }
        label: { type: string }
        status: { type: string }
        confidence_score: { type: number, minimum: 0, maximum: 1 }
        content_layer:
          type: string
          enum: [biomedical, education, learning]
        external_ids: { type: object, additionalProperties: true }
...
    ExplainResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            question: { type: string }
            answer_summary: { type: string }
            reasoning_chain: { type: array, items: { type: object } }
            confidence: { type: number }
            evidence_level: { type: string }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...
    SearchResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              entity: { $ref: "#/components/schemas/EntitySummary" }
              score: { type: number }
              snippet: { type: string }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }